Support multiple VMs; change interface to match other cloud plugins.
